requirements.txt
# Core dependencies
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
biopython>=1.83
chromadb>=0.4.22
fastapi>=0.110.0
google-generativeai>=0.3.2
httpx>=0.27.0
langchain>=0.1.12
langchain-community>=0.0.27
langchain-core>=0.1.27
langchain-google-genai>=0.0.11
langgraph>=0.0.27
pydantic>=2.6.3
python-dotenv>=1.0.1
requests>=2.31.0
sentence-transformers>=2.5.1
uvicorn>=0.27.1

# Development dependencies
black>=24.2.0
isort>=5.13.2
mypy>=1.8.0
pytest>=8.0.2

# Optional dependencies
exa-py>=0.1.0  # For EXA search integration

# Note: bioconductor is not available for Python 3.13
# Consider using alternative bioinformatics packages or downgrading Python version if bioconductor is required

# Smithery integration
smithery-sdk>=0.1.0  # Replace with actual package name when available
